Melhorias Gerais
1° - Removido a opção de excluir pessoa e excluir veículos
2° - Criado a opção de desativar a pessoa.
    2.1 - Pessoas desativadas não aparecem na pesquisa de pessoas por padrão, criado botão para mostrar as pessoas desativadas. Sendo possível ativar elas novamente.
    2.2 - Veículos que tem a pessoa desativada não aparece na consulta de veículo.
3° - Cadastro de pessoa
    3.1 - Agora permite trocar o tipo de preço, só tem que saber o ID.
    3.2 - Agora permite trocar o telefone, e o endereço da pessoa, e não permite mais editar o CPF

+++ src/app/pessoa/conponents/listarPessoa.tsx
@@ -11,39 +11,43 @@ interface Registro {
   descricao: string;
   valor: number;
   valores_id: number;
+  pago: boolean;
+  ativo: boolean;
 }
 
 export function TabelaRegistros() {
   const [registros, setRegistros] = useState<Registro[]>([]);
   const [editando, setEditando] = useState<number | null>(null);
   const [valoresEditados, setValoresEditados] = useState<Partial<Registro>>({});
+  const [mostrarInativos, setMostrarInativos] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   const fetchRegistros = async () => {
+    setLoading(true);
     try {
-      const response = await fetch("http://localhost:3000/api/cadastroPessoa");
+      // Adiciona o parâmetro `inativos` à URL com base no estado de `mostrarInativos`
+      const url = `http://localhost:3000/api/cadastroPessoa${mostrarInativos ? "?inativos=true" : ""}`;
+      const response = await fetch(url);
+      if (!response.ok) throw new Error("Erro ao buscar registros");
+
       const data = await response.json();
-      setRegistros(data);
+      // Adaptando o campo `situacaoPessoa` para `ativo`
+      const registrosAdaptados = data.map((registro: any) => ({
+        ...registro,
+        ativo: registro.situacaoPessoa,
+      }));
+      setRegistros(registrosAdaptados);
     } catch (error) {
       console.error("Erro ao buscar registros:", error);
+      alert("Erro ao buscar registros");
+    } finally {
+      setLoading(false);
     }
   };
 
   useEffect(() => {
     fetchRegistros();
-  }, []);
-
-  const formatarTelefone = (telefone: string) => {
-    const apenasNumeros = telefone.replace(/\D/g, "");
-    return apenasNumeros.replace(
-      /(\d{2})(\d{1})(\d{4})(\d{4})/,
-      "($1) $2 $3-$4",
-    );
-  };
-
-  const formatarCPF = (cpf: string) => {
-    const apenasNumeros = cpf.replace(/\D/g, "");
-    return apenasNumeros.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, "$1.$2.$3-$4");
-  };
+  }, [mostrarInativos]);
 
   const iniciarEdicao = (registro: Registro) => {
     setEditando(registro.id);
@@ -62,55 +66,72 @@ export function TabelaRegistros() {
   };
 
   const salvarEdicao = async (id: number) => {
-    const novoRegistro = {
-      id: id,
-      ...valoresEditados,
-    };
-
     try {
+      setLoading(true);
       const response = await fetch("http://localhost:3000/api/cadastroPessoa", {
         method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(novoRegistro),
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ id, ...valoresEditados }),
       });
+      if (!response.ok) throw new Error("Erro ao atualizar o registro");
 
-      if (response.ok) {
-        alert("Registro atualizado com sucesso!");
-        fetchRegistros(); // Recarrega a lista de registros após o update
-        console.log("Cheguei");
-      } else {
-        alert("Erro ao atualizar o registro.");
-      }
+      alert("Registro atualizado com sucesso!");
+      fetchRegistros();
     } catch (error) {
       console.error("Erro ao enviar atualização:", error);
-      alert("Erro ao atualizar o registro.");
+      alert("Erro ao atualizar o registro");
     } finally {
       setEditando(null);
       setValoresEditados({});
+      setLoading(false);
     }
   };
 
   const excluirRegistro = async (id: number) => {
+    if (!confirm("Deseja realmente excluir o registro?")) return;
     try {
+      setLoading(true);
       const response = await fetch("http://localhost:3000/api/cadastroPessoa", {
         method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
+        headers: { "Content-Type": "application/json" },
         body: JSON.stringify({ id }),
       });
+      if (!response.ok) throw new Error("Erro ao excluir o registro");
 
-      if (response.ok) {
-        alert("Registro excluído com sucesso!");
-        fetchRegistros(); // Recarregar a lista de registros
-      } else {
-        console.error("Erro ao excluir registro:", await response.json());
-      }
+      alert("Registro excluído com sucesso!");
+      fetchRegistros();
     } catch (error) {
       console.error("Erro ao excluir registro:", error);
+      alert("Erro ao excluir registro");
+    } finally {
+      setLoading(false);
     }
+  };
+
+  const atualizarStatusAtivo = async (id: number, ativo: boolean) => {
+    try {
+      const response = await fetch("http://localhost:3000/api/cadastroPessoa", {
+        method: "PATCH",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ id, toggleSituacao: true }),
+      });
+      if (!response.ok) throw new Error("Erro ao atualizar status de ativação");
+
+      alert(`Registro ${ativo ? "ativado" : "desativado"} com sucesso!`);
+      fetchRegistros();
+    } catch (error) {
+      console.error("Erro ao atualizar ativação:", error);
+      alert("Erro ao atualizar ativação");
+    }
+  };
+
+  const handleAtivarDesativarClick = (registro: Registro) => {
+    const novoStatus = !registro.ativo;
+    atualizarStatusAtivo(registro.id, novoStatus);
+  };
+
+  const handleToggleMostrarInativos = () => {
+    setMostrarInativos((prev) => !prev);
   };
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -120,6 +141,9 @@ export function TabelaRegistros() {
 
   return (
     <div style={styles.tableContainer}>
+      <button onClick={handleToggleMostrarInativos} style={styles.filterButton}>
+        {mostrarInativos ? "Mostrar Ativos" : "Mostrar Inativos"}
+      </button>
       <table style={styles.table}>
         <thead>
           <tr>
@@ -127,124 +151,130 @@ export function TabelaRegistros() {
             <th style={styles.headerCell}>CPF</th>
             <th style={styles.headerCell}>Telefone</th>
             <th style={styles.headerCell}>Endereço</th>
+            <th style={styles.headerCell}>ID tipo valor</th>
             <th style={styles.headerCell}>Tipo Valor</th>
             <th style={styles.headerCell}>Valor</th>
             <th style={styles.headerCell}>Ações</th>
           </tr>
         </thead>
         <tbody>
-          {Array.isArray(registros) && registros.length > 0 ? (
-            registros.map((registro) => (
-              <tr
-                key={registro.id}
-                style={editando === registro.id ? styles.editRow : {}}
-              >
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <input
-                      type="text"
-                      name="nome"
-                      value={valoresEditados.nome || ""}
-                      onChange={handleInputChange}
-                      style={styles.input}
-                    />
-                  ) : (
-                    registro.nome
-                  )}
-                </td>
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <input
-                      type="text"
-                      name="cpf"
-                      value={valoresEditados.cpf || ""}
-                      onChange={handleInputChange}
-                      style={styles.input}
-                    />
-                  ) : (
-                    formatarCPF(registro.cpf)
-                  )}
-                </td>
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <input
-                      type="text"
-                      name="telefone"
-                      value={valoresEditados.telefone || ""}
-                      onChange={handleInputChange}
-                      style={styles.input}
-                    />
-                  ) : (
-                    formatarTelefone(registro.telefone)
-                  )}
-                </td>
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <input
-                      type="text"
-                      name="Endereço"
-                      value={valoresEditados.endereco || ""}
-                      onChange={handleInputChange}
-                      style={styles.input}
-                    />
-                  ) : (
-                    registro.endereco
-                  )}
-                </td>
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <input
-                      type="text"
-                      name="valores_id"
-                      value={valoresEditados.valores_id || ""}
-                      onChange={handleInputChange}
-                      style={styles.input}
-                    />
-                  ) : (
-                    registro.valores_id
-                  )}
-                </td>
-                <td style={styles.cell}>
-                  {registro.valor != null ? registro.valor : "Não informado"}
-                </td>
-
-                <td style={styles.cell}>
-                  {editando === registro.id ? (
-                    <>
-                      <button
-                        onClick={() => salvarEdicao(registro.id)}
-                        style={styles.saveButton}
-                      >
-                        Salvar
-                      </button>
-                      <button
-                        onClick={cancelarEdicao}
-                        style={styles.cancelButton}
-                      >
-                        Cancelar
-                      </button>
-                    </>
-                  ) : (
-                    <>
-                      <button
-                        onClick={() => iniciarEdicao(registro)}
-                        style={styles.editButton}
-                      >
-                        Editar
-                      </button>
-                      <button
-                        onClick={() => excluirRegistro(registro.id)}
-                        style={styles.deleteButton}
-                      >
-                        Excluir
-                      </button>
-                    </>
-                  )}
-                </td>
-              </tr>
-            ))
+          {loading ? (
+            <tr>
+              <td colSpan={7} style={styles.cell}>
+                Carregando...
+              </td>
+            </tr>
+          ) : registros.length > 0 ? (
+            registros
+              .filter((registro) => registro.ativo === !mostrarInativos)
+              .map((registro) => (
+                <tr
+                  key={registro.id}
+                  style={editando === registro.id ? styles.editRow : {}}
+                >
+                  <td style={styles.cell}>
+                    {editando === registro.id ? (
+                      <input
+                        type="text"
+                        name="nome"
+                        value={valoresEditados.nome || ""}
+                        onChange={handleInputChange}
+                        style={styles.input}
+                      />
+                    ) : (
+                      registro.nome
+                    )}
+                  </td>
+                  <td style={styles.cell}>{registro.cpf}</td>
+                  <td style={styles.cell}>
+                    {editando === registro.id ? (
+                      <input
+                        type="text"
+                        name="telefone"
+                        value={valoresEditados.telefone || ""}
+                        onChange={handleInputChange}
+                        style={styles.input}
+                      />
+                    ) : (
+                      registro.telefone
+                    )}
+                  </td>
+                  <td style={styles.cell}>
+                    {editando === registro.id ? (
+                      <input
+                        type="text"
+                        name="endereco"
+                        value={valoresEditados.endereco || ""}
+                        onChange={handleInputChange}
+                        style={styles.input}
+                      />
+                    ) : (
+                      registro.endereco
+                    )}
+                  </td>
+                  <td style={styles.cell}>
+                    {editando === registro.id ? (
+                      <input
+                        type="text"
+                        name="valores_id"
+                        value={valoresEditados.valores_id || ""}
+                        onChange={handleInputChange}
+                        style={styles.input}
+                      />
+                    ) : (
+                      registro.valores_id
+                    )}
+                  </td>
+                  <td style={styles.cell}>{registro.descricao}</td>
+                  <td style={styles.cell}>
+                    {registro.valor != null ? registro.valor : "Não informado"}
+                  </td>
+                  <td style={styles.cell}>
+                    {editando === registro.id ? (
+                      <>
+                        <button
+                          onClick={() => salvarEdicao(registro.id)}
+                          style={styles.saveButton}
+                        >
+                          Salvar
+                        </button>
+                        <button
+                          onClick={cancelarEdicao}
+                          style={styles.cancelButton}
+                        >
+                          Cancelar
+                        </button>
+                      </>
+                    ) : (
+                      <>
+                        <button
+                          onClick={() => iniciarEdicao(registro)}
+                          style={styles.editButton}
+                        >
+                          Editar
+                        </button>
+                        <button
+                          onClick={() => handleAtivarDesativarClick(registro)}
+                          style={{
+                            ...styles.activationButton,
+                            backgroundColor: registro.ativo
+                              ? "#FF9800"
+                              : "#4CAF50",
+                          }}
+                        >
+                          {registro.ativo ? "Desativar" : "Ativar"}
+                        </button>
+                      </>
+                    )}
+                  </td>
+                </tr>
+              ))
           ) : (
-            <tr></tr>
+            <tr>
+              <td colSpan={7} style={styles.cell}>
+                Nenhum registro encontrado
+              </td>
+            </tr>
           )}
         </tbody>
       </table>
@@ -253,8 +283,15 @@ export function TabelaRegistros() {
 }
 
 const styles: { [key: string]: React.CSSProperties } = {
-  tableContainer: {
-    width: "100%",
+  tableContainer: { width: "100%", padding: "20px" },
+  filterButton: {
+    marginBottom: "15px",
+    padding: "10px 15px",
+    borderRadius: "4px",
+    border: "none",
+    cursor: "pointer",
+    backgroundColor: "#007BFF",
+    color: "#FFFFFF",
   },
   table: {
     width: "100%",
@@ -263,7 +300,7 @@ const styles: { [key: string]: React.CSSProperties } = {
   },
   headerCell: {
     padding: "10px 15px",
-    fontWeight: "normal",
+    fontWeight: "bold",
     borderBottom: "1px solid #ddd",
     textAlign: "left",
   },
@@ -273,9 +310,7 @@ const styles: { [key: string]: React.CSSProperties } = {
     borderBottom: "1px solid #ddd",
     textAlign: "left",
   },
-  editRow: {
-    backgroundColor: "#e0f7fa",
-  },
+  editRow: { backgroundColor: "#e0f7fa" },
   input: {
     width: "100%",
     padding: "5px",
@@ -309,12 +344,12 @@ const styles: { [key: string]: React.CSSProperties } = {
     cursor: "pointer",
     marginRight: "5px",
   },
-  deleteButton: {
-    backgroundColor: "#f44336",
+  activationButton: {
     color: "white",
     padding: "5px 10px",
     border: "none",
     borderRadius: "4px",
     cursor: "pointer",
+    marginRight: "5px",
   },
 };
